Fixed an minor error with the naps2 pkg. Read commit description.
This commit fixes a minor yet important error in regards to the OCR
functinality of NAPS2. If the tesserract binary did not have executable
permissions, it would cause the OCR process to not work at all. Which
results in a PDF with no OCR. This commit aimes to directly fix that
issue.

Signed-off-by: kris3713 <07e5297d5b@c0x0.com>